docconverter/requirements.txt
accelerate==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
attrs==25.4.0
beautifulsoup4==4.14.3
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
cryptography==44.0.1
defusedxml==0.7.1
Deprecated==1.3.1
dill==0.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
docling-core==2.68.0
docling-ibm-models==3.11.0
docling-parse==5.5.0
et_xmlfile==2.0.0
Faker==40.8.0
filelock==3.25.0
filetype==1.2.0
fire==0.7.0
fonttools==4.56.0
fsspec==2026.2.0
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-firestore==2.20.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
gunicorn==23.0.0
httplib2==0.22.0
huggingface_hub==0.36.2
idna==3.10
img2pdf==0.6.3
Jinja2==3.1.6
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
latex2mathml==3.78.1
lxml==5.3.1
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.0
multiprocess==0.70.19
networkx==3.4.2
numpy==2.2.3
omegaconf==2.3.0
opencv-python==4.13.0.92
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.3.3
pdf2docx==0.5.8
pikepdf==10.3.0
pillow==11.1.0
pluggy==1.6.0
polyfactory==3.3.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==7.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.4.0
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pylatexenc==2.10
PyMuPDF==1.25.3
pyparsing==3.2.1
pypdfium2==5.6.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.2.2
python-pptx==1.0.2
pytz==2026.1.post1
PyYAML==6.0.3
rapidocr==3.7.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.3
rich==14.3.3
rpds-py==0.30.0
rsa==4.9
rtree==1.4.1
safetensors==0.7.0
scipy==1.15.3
semchunk==2.2.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
SpeechRecognition==3.14.1
sqlparse==0.5.3
sympy==1.14.0
tabulate==0.9.0
termcolor==2.5.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.21.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
wrapt==2.1.2
xlsxwriter==3.2.9
